\name{examples.ncv}

\alias{examples.ncv}
\alias{foo.ncv}
\alias{ex.pavel.ncv}
\alias{ex.cf5.1.ncv}
\alias{ex.cf5.2.ncv}
\alias{ex.cf5.6.ncv}
\alias{ex.cf7.2.ncv}

\title{Examples}

\description{Executes the example functions.}

\usage{example()}

\arguments{
}

\details{Executes the example functions. The examples provide
further illustration of how to use this package.
To see the code enter the example name without parenthesis,
e.g. \command{foo.ncv}.}
}

\value{List of example functions.}

\author{Juerg Schmidli}

\examples{
foo.ncv()		# default example, used for the help pages
ex.cf5.1.ncv()		# example 5.1 from the CF-conventions
ex.cf5.2.ncv()		# example 5.2 from the CF-conventions
ex.cf5.6.ncv()		# example 5.6 from the CF-conventions
ex.cf7.2.ncv()		# example 7.2 from the CF-conventions
}

\keyword{file}
